# reactapp/src/services/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "classroom",
    "version": "1.0"
  },
  "paths": {
    "/api/Hello": {
      "get": {
        "tags": [
          "Hello"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Hello"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Hello"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Hello"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Hello"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Hello"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Hello"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Hello"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Hello"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hello"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hello"
                }
              }
            }
          }
        }
      }
    },
    "/api/Hello/{id}": {
      "get": {
        "tags": [
          "Hello"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Hello"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hello"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hello"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Hello"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Hello"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Hello"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Hello"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Hello"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/WeatherForecast": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "operationId": "GetWeatherForecast",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Hello": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WeatherForecast": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "temperatureC": {
            "type": "integer",
            "format": "int32"
          },
          "temperatureF": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "summary": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}